feat: Jumbotron now auto-rotates between projects.
Beginning with the initial project for the page, the jumbotron will rotate across all project images. It should also change to the first project of a subject area when the corresponding button is pressed. It should also interrupt the interval and resume from the currently-active jumbotron project any time the jumbotron is changed through the other triggers.