feat: Experiment with interrupting the jumbotron interval
In the process, highlighting the need for a consistent MVC framework of my own design or perhaps Angular or React. Also, the quirky behavior of the system at this point highlights the need to implement the accordion and carousel idea, along with a solid javascript front end framework. Such an overhaul should showcase my design and technical skills as they have progressed to this point as well as improving the functionality, aesthetic, and user experience of my site